Refactor: TextField -> SearchBar로 변경
- SearchTextField를 UITextField에서 UISearchBar로 변경
- SearchTextField -> DestinationSearchBar로 이름 뱐걍

#5
